==> Address/prajwal/Feb 20/mn 1850/lib/01_doubt-Copy1.ipynb <==
import pandas as pd
import numpy as np

#df = pd.read_excel('/Users/prajwal/Desktop/Columbia/C4SR/mn_1880/not so confident.xlsx')

#df = pd.read_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract output/MN_1880_final_output_tesseract.csv')
df = pd.read_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/MN_1850 tesseract spl removed.csv')


#df_slice1 = df.loc[df['filter1']==1]

df_slice1 = df

try:
    df_slice1 = df_slice1.rename({'clean address': 'clean_address'}, axis=1)
except:
    pass
try:
    df = df.rename({'clean address': 'clean_address'}, axis=1)
except:
    pass

#slice1_addresses = list(dict(df_slice1['clean_address']).values()) #was initially uncommented



#slice1_addresses = list(dict(df_slice1['clean address']).values())

slice1_addresses = list(dict(df_slice1['Full Address']).values())

modified_clean_address = df['Full Address'].str.replace('Ã¢Â\x80Â\x99',"'").str.replace('Ã¢ÂÂ',"'").str.replace("Ã¢Â\x80Â\x98","").str.replace("Ã¢Â\x80Â\x9c","").str.replace("Ã¢Â\x80Â\x9d","")


"""


#modified_clean_address = df['clean_address'].str.replace('Ã¢Â\x80Â\x99',"'").str.replace('Ã¢ÂÂ',"'").str.replace("Ã¢Â\x80Â\x98","").str.replace("Ã¢Â\x80Â\x9c","").str.replace("Ã¢Â\x80Â\x9d","")
modified_clean_address = modified_clean_address.str.replace("hN", "h N")
print(modified_clean_address[0])
modified_clean_address = modified_clean_address.str.replace(".5 ", " ")
print(modified_clean_address[0])
modified_clean_address = modified_clean_address.str.replace(",h ", ", h ")
print(modified_clean_address[0])
modified_clean_address = modified_clean_address.str.replace(",h", ", h ")
print(modified_clean_address[0])
modified_clean_address = modified_clean_address.str.replace(", h.", ", h. ")
print(modified_clean_address[0])
modified_clean_address = modified_clean_address.str.replace(", h .", ", h. ")
print(modified_clean_address[0])
modified_clean_address = modified_clean_address.str.replace("H udson", "Hudson")
modified_clean_address = modified_clean_address.str.replace("'wav", "'way")
modified_clean_address = modified_clean_address.str.replace("B'way", "Broadway")
modified_clean_address = modified_clean_address.str.replace("B way", "Broadway")
modified_clean_address = modified_clean_address.str.replace("B'kly n", "B'klyn")
modified_clean_address = modified_clean_address.str.replace(",n", " n")
modified_clean_address = modified_clean_address.str.replace("W.h", " Wh")
modified_clean_address = modified_clean_address.str.replace("h-", " h ")
modified_clean_address = modified_clean_address.str.replace("h-", " h ")
modified_clean_address = modified_clean_address.str.replace("-h.", "h.")
#modified_clean_address = modified_clean_address.str.replace("H()", "h.")
#modified_clean_address = modified_clean_address.str.replace("*", " ")
#modified_clean_address = modified_clean_address.str.replace("h r.", "h r")

#modified_clean_address = modified_clean_address.str.replace(".h", ", h.")

#modified_clean_address = modified_clean_address.str.replace("^h", "  h")
#modified_clean_address = modified_clean_address.str.replace("h*", "  h ")

#modified_clean_address = modified_clean_address.str.replace(". .", " ")
"""


for i in range(0,len(modified_clean_address)):
    if str(modified_clean_address[i]) == 'nan':
        print(i, modified_clean_address[i])
        modified_clean_address[i] = ''


for i in range(0,len(modified_clean_address)):
        if modified_clean_address[i].find("E.") >= 0:
            modified_clean_address[i] = modified_clean_address[i].replace('E.','E ').replace('  ', ' ')

for i in range(0,len(modified_clean_address)):
    if modified_clean_address[i].find("W.") >= 0:
        modified_clean_address[i] = modified_clean_address[i].replace('W.','W ').replace('  ', ' ')

for i in range(0,len(modified_clean_address)):
    if modified_clean_address[i].find(".h") >= 0:
        modified_clean_address[i] = modified_clean_address[i].replace('.h',', h. ').replace(',,',',').replace('  ',' ')

        

for i in range(0,len(modified_clean_address)):
    if modified_clean_address[i].find(",c") >= 0:
        modified_clean_address[i] = modified_clean_address[i].replace(',c',' c')

for i in range(0,len(modified_clean_address)):
    if modified_clean_address[i].find(" c. ") >= 0:
        modified_clean_address[i] = modified_clean_address[i].replace('c.','c')

for i in range(0,len(modified_clean_address)):
    if modified_clean_address[i].find(" n. ") >= 0:
        modified_clean_address[i] = modified_clean_address[i].replace(' n. ',' n ')

for i in range(0,len(modified_clean_address)):
    if modified_clean_address[i].find("h r ") >=0:
        modified_clean_address[i] = modified_clean_address[i].replace('h r','h. rear').replace('rear.','rear ')

for i in range(0,len(modified_clean_address)):
    if modified_clean_address[i].find("h. r ") >=0:
        modified_clean_address[i] = modified_clean_address[i].replace('h. r','h. rear').replace('rear.','rear ')
        
        

for i in range(0,len(modified_clean_address)):
    if modified_clean_address[i][0:2].find("h'") >=0:
        modified_clean_address[i] = modified_clean_address[i].replace("h'","h. ")

for i in range(0,len(modified_clean_address)):
    if modified_clean_address[i][0:2].find("th") >=0:
        modified_clean_address[i] = modified_clean_address[i].replace("th","h")

modified_clean_address[0]

#re.findall(r'. 1', '.1')

import re

modified_clean_address4 = []
intersections_n_1 = []
intersections_n_2 = []

intersections_c_1 = []
intersections_c_2 = []

for i in range(0,len(modified_clean_address)):
    temp = modified_clean_address[i]
    
    
    
    
    """
    string_with_th = re.findall(r'[0-9]+th', temp)
    for j in range (0, len(string_with_th)):
        temp = temp.replace(string_with_th[j], string_with_th[j].replace('th',''))
    
    string_with_d = re.findall(r'[0-9]+d', temp)
    for j in range (0, len(string_with_d)):
        temp = temp.replace(string_with_d[j], string_with_d[j].replace('d',''))
    
    string_with_rd = re.findall(r'[0-9]+rd', temp)
    for j in range (0, len(string_with_rd)):
        temp = temp.replace(string_with_rd[j], string_with_rd[j].replace('rd',''))
    
    string_with_st = re.findall(r'[0-9]+st', temp)
    for j in range (0, len(string_with_st)):
        temp = temp.replace(string_with_st[j], string_with_st[j].replace('st',''))
    
    string_with_h_dot = re.findall(r'h.+[0-9]', temp)
    for j in range (0, len(string_with_h_dot)):
        temp = temp.replace(string_with_h_dot[j], string_with_h_dot[j].replace('h.','h. ').replace('  ',' '))
    
    string_with_h = re.findall(r'h+[0-9]', temp)
    for j in range (0, len(string_with_h)):
        temp = temp.replace(string_with_h[j], string_with_h[j].replace('h','h ').replace('  ',' '))
        
    string_with_fullstop = re.findall(r'. +[0-9]', temp)
    for j in range (0, len(string_with_fullstop)):
        temp = temp.replace(string_with_fullstop[j], string_with_fullstop[j].replace('.','. h ').replace('  ',' '))
    
    string_with_comma = re.findall(r', +[0-9]', temp)
    for j in range (0, len(string_with_comma)):
        temp = temp.replace(string_with_comma[j], string_with_comma[j].replace(',',', h ').replace('  ',' '))
    
    string_with_bracket = re.findall(r'\) +[0-9]', temp)
    for j in range (0, len(string_with_bracket)):
        temp = temp.replace(string_with_bracket[j], string_with_bracket[j].replace(')',') h ').replace('  ',' '))
    
    string_with_dash = re.findall(r'-+[0-9]', temp)
    for j in range (0, len(string_with_dash)):
        temp = temp.replace(string_with_dash[j], string_with_dash[j].replace('-','').replace('  ',' '))
    
    string_with_fullstop_number = re.findall(r'.+[0-9]', temp)
    for j in range (0, len(string_with_fullstop_number)):
        temp = temp.replace(string_with_fullstop_number[j], string_with_fullstop_number[j].replace('.','').replace('  ',' '))
    
    
    string_with_S_1 = re.findall(r'[0-9]+S', temp)
    string_with_S_2 = re.findall(r'S+[0-9]', temp)
    string_with_S = string_with_S_1 + string_with_S_2
    for j in range (0, len(string_with_S)):
        #print(i,string_with_S)
        temp = temp.replace(string_with_S[j], string_with_S[j].replace('S','5'))
        #print(temp)
    
    
    """
    
    
    
    try:
        location_of_h_1 = temp.split().index('h.')
    except:
        location_of_h_1 = -1
    try:
        location_of_h_2 = temp.split().index('h')
    except:
        location_of_h_2 = -1
    #print(location_of_h_1, location_of_h_2)

    location_of_h = max(location_of_h_1, location_of_h_2)
    """
    if location_of_h_1>0:
        if location_of_h_2<0:
            location_of_h = location_of_h_1
        else:
            #print(i,'in exception1')
            #print(location_of_h_1, location_of_h_2)
            location_of_h = max(location_of_h_1, location_of_h_2)
    
    else:
        if location_of_h_2 > 0:
            location_of_h = location_of_h_2
        else:
            print(i,'in exception2')
            print(location_of_h_1, location_of_h_2)
            location_of_h = -1
    """
    temp = temp.replace('^','')
    
    
    
    """
    temp = temp.replace('ft.','')
    temp = temp.replace('ft','')
    temp = temp.replace('Ft.','')
    """
    
    undesired_starts = ['.',' .',',',' ,','-',' -','- ',', ','. ',"'","' "," '","''",'&','*','& ','* ',
                       ' &',' *']
    for i in range(0,len(undesired_starts)):
        if temp[0:2].find(undesired_starts[i])>=0:
            temp = temp[0:2].replace(undesired_starts[i],'') + temp[2:]
    
    words_after_h = temp.split()[location_of_h+1:]
    try:
        location_of_n = words_after_h.index('n')
        intersection_n_street1 = ' '.join(words_after_h[0:location_of_n])
        intersection_n_street2 = ' '.join(words_after_h[location_of_n+1:])

        intersections_n_1.append(intersection_n_street1)
        intersections_n_2.append(intersection_n_street2)
    except:
        intersections_n_1.append('')
        intersections_n_2.append('')
    
    try:
        location_of_c = words_after_h.index('c')
        intersection_c_street1 = ' '.join(words_after_h[0:location_of_c])
        intersection_c_street2 = ' '.join(words_after_h[location_of_c+1:])

        intersections_c_1.append(intersection_c_street1)
        intersections_c_2.append(intersection_c_street2)
    except:
        intersections_c_1.append('')
        intersections_c_2.append('')
    
    modified_clean_address4.append(temp)

#for i in range(0,len(intersections_c_1)):
    #print(i, intersections_c_1[i] + ' and ' + intersections_c_2[i])

for i in range(0,len(modified_clean_address4)):
    if modified_clean_address4[i].find(".h") >= 0:
        print(i, modified_clean_address4[i])

#for i in range(0,len(modified_clean_address)):
    #if modified_clean_address[i].find(" n ") >0:
        #print(i, modified_clean_address[i])

df['modified_clean_address'] = modified_clean_address4

df['n_street1'] = intersections_n_1
df['n_street2'] = intersections_n_2
df['c_street1'] = intersections_c_1
df['c_street2'] = intersections_c_2

#df.head()

#df.to_excel('doubt/not so confident2.xlsx')

df_modified_addresses = pd.DataFrame(modified_clean_address4, columns=['clean_address'])

#pd.DataFrame(modified_clean_address2).to_csv('doubt/doubt2.txt', header=None,  sep=' ', mode='a')

temp_addr = []
for i in range(0,len(df_modified_addresses)):
    temp_addr.append(str(i)+" "+df_modified_addresses['clean_address'][i])
#np.savetxt('doubt/not so confident addr.txt', temp_addr, fmt='%s')


#np.savetxt('doubt/not so confident addr.txt', temp_addr, fmt='%s')


#df.head()

#df.columns

#df.head()

df['n_street1'] = df['n_street1'].replace('',np.nan)
df['n_street2'] = df['n_street2'].replace('',np.nan)
df['c_street1'] = df['c_street1'].replace('',np.nan)
df['c_street2'] = df['c_street2'].replace('',np.nan)

#df['filter11'] = np.where(df['n_street1'].isna() & df['n_street2'].isna() & df['c_street1'].isna() & df['c_street2'].isna() & df['new h house number'].isna() & df['new h street name'].isna(), 1, 0)
#was initially uncommented

df['filter11'] = np.where(df['n_street1'].isna() & df['n_street2'].isna() & df['c_street1'].isna() & df['c_street2'].isna(), 1, 0)

df_filter11 = df.loc[df['filter11'] == 1]
df_filter11 = df_filter11.reset_index()

#df_filter11['modified_clean_address']

modified_clean_address5 = []
for i in range(0,len(df_filter11)):
    temp = df_filter11['modified_clean_address'][i]
    
    string_with_S = re.findall(r'[0-9]+S', temp)
    for i in range (0, len(string_with_S)):
        temp = temp.replace(string_with_S[i], string_with_S[i].replace('S','5'))
    modified_clean_address5.append(temp)   

#for i in range(0,len(modified_clean_address5)):
    #print(i, modified_clean_address5[i])

#for i in range(0,len(modified_clean_address5)):
    #if modified_clean_address5[i].find("h. ") <0:
        #print(i, modified_clean_address5[i])

df['modified_clean_address'][10:16]

df['modified_clean_address'] = df['modified_clean_address'].str.replace('rearear','rear')
df['modified_clean_address'] = df['modified_clean_address'].str.replace('rearefused','refused')
df['modified_clean_address'] = df['modified_clean_address'].str.replace('h. r ','h. rear ')


df

#df.to_excel('/Users/prajwal/Desktop/Columbia/C4SR/mn_1880/not so confident-mn1880-v7.xlsx')

df.to_excel('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/mn-1850-2.xlsx')
