==> Address/prajwal/Feb 20/mn 1850/lib/07_before address builder-Copy1.ipynb <==
import pandas as pd
import numpy as np
import re

mn_or_bk = 'mn' #bk or mn
year = '1850'
#doubttype = 'not so confident' #doubt or not so confident
#doubttype = 'not so confident' #doubt or not so confident

#df = pd.read_csv('/Users/prajwal/Desktop/Columbia/C4SR/' + mn_or_bk + '_' + year + '/after cleaning/' + mn_or_bk + '_' + year + '_after_cleaning_' + doubttype + '4.csv')

#df = pd.read_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract output/run2/after cleaning/bk-1850-3-s-combine-clean.csv')
df = pd.read_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/mn-1850-3-s-combine-clean.csv')




df.columns = ['index', 'Full Address', 'H House Number','H Street Name', 
              'c_street1', 'c_street2', 'n_street1', 'n_street2', 'H City',
              'H Address', 'Address 2', 'Address 2 House Number', 'Address 2 Street Name',
              'Address 2 City'
             ]

"""
df.columns = ['index', 'Full Address', 'H House Number', 
              'H Street Name', 'c_street1', 'c_street2', 'n_street1', 'n_street2', 
              'new h city', 'new h house number', 'new h street name', 'unidentified h info']

df.columns = ['index', 'n_street1', 'n_street2','c_street1', 'c_street2', 'Full Address', 
              'H House Number', 
              'H Street Name', 
              'new h house number', 'new h street name','new h city', 'unidentified h info']
"""

df = df.drop(['index'], axis=1)

#df.loc[pd.isna(df['H House Number']) == True and pd.isna(df['Address 2']) == False]

df.head(100)

# Combines data from various columns into two: hn_1 (for house number) and best_match (for street name)

hn_1 = {}
best_match = {}

for i in range(0,len(df)):
    hn_1[i] = df['H House Number'][i]
    best_match[i] = df['H Street Name'][i]

for i in range(0,len(df)):
#for i in range(35,36):
    #print('begin',hn_1[i],best_match[i])
    
    #handle corner cases
    if pd.isna(df['c_street1'][i]) == False or pd.isna(df['c_street2'][i]) == False:
        #print('1',i)
        try:
            best_match[i] = str(df['c_street1'][i]) + ' & ' + str(df['c_street2'][i])
        except:
            if pd.isna(df['c_street1'][i]) == False:
                best_match[i] = str(df['c_street1'])
            else:
                best_match[i] = str(df['c_street2'])
                
    #handle near cases
    elif pd.isna(df['n_street1'][i]) == False or pd.isna(df['n_street2'][i]) == False:
        #print('2',i)
        try:
            best_match[i] = str(df['n_street1'][i]) + ' & ' + str(df['n_street2'][i])
        except:
            if pd.isna(df['n_street1'][i]) == False:
                best_match[i] = str(df['n_street1'])
            else:
                best_match[i] = str(df['n_street2'])
    
    #if corner and near cases don't have any data, use H House Number and H Street Name
    elif pd.isna(hn_1[i]) == True or pd.isna(best_match[i]) == True:
        #print('3',i)
        #try:
            #hn_1[i] = df['H House Number'][i]
        #except:
            #pass
        try:
            best_match[i] = df['H City'][i]
        except:
            pass
        
        
    """
    #if everything above didn't work, use new h house number and new h street name
    elif pd.isna(hn_1[i]) == True or pd.isna(best_match[i]) == True:
        #print('4',i)
        try:
            hn_1[i] = df['new h house number'][i]
        except:
            pass
        try:
            best_match[i] = df['new h street name'][i]
        except:
            pass
    
    #if everything above still didn't work, use new h city for best_match
    elif pd.isna(df['new h city'][i]) == False and pd.isna(best_match[i]) == True:
        #print('5',i)
        best_match[i] = df['new h city'][i]
    
    # return a blank house number and Full Address for best_match if nothing above worked
    else:
        if pd.isna(hn_1[i]) == True and pd.isna(best_match[i]) == True:
            #print('6',i)
            hn_1[i] = ''
            best_match[i] = df['Full Address'][i]
            
    """
    
    """
    elif pd.isna(df['H House Number'][i]) == False or pd.isna(df['H Street Name'][i]) == False:
        print('3',i)
        try:
            hn_1[i] = df['H House Number'][i]
        except:
            pass
        try:
            best_match[i] = df['H Street Name'][i]
        except:
            pass
    
    elif pd.isna(df['new h house number'][i]) == False or pd.isna(df['new h street name'][i]) == False:
        print('4',i)
        try:
            hn_1[i] = df['new h house number'][i]
        except:
            pass
        try:
            best_match[i] = df['new h street name'][i]
        except:
            pass
    
    elif pd.isna(df['new h city'][i]) == False:
        print('5',i)
        best_match[i] = df['new h city'][i]
    
    else:
        print('6',i)
        hn_1[i] = ''
        best_match[i] = df['Full Address'][i]
    """
    #if pd.isna(df['H House Number'][i]) == True and pd.isna(df['new h house number'][i]) == True:
        #print('7',i)
        #hn_1[i] = ''
    
    if pd.isna(hn_1[i]) == True and pd.isna(best_match[i]) == True:
        #print('7',i)
        hn_1[i] = ''
        best_match[i] = df['H Address'][i]



df2 = pd.DataFrame([list(hn_1.values()), list(best_match.values())])

df2 = df2.T


df2.columns = ['hn_1','best_match']

"""
def find_number(text):
    num = re.findall(r'[0-9]+',str(text))
    return " ".join(num)
df2['hn_1']=df2['hn_1'].apply(lambda x: find_number(x))
"""

df2

#df2.to_csv('/Users/prajwal/Desktop/Columbia/C4SR/' + mn_or_bk + '_' + year + '/for address builder/' + mn_or_bk + '_' + year + '_' + doubttype + '_ready4.csv')
df2.to_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/mn-1850-for-address-builder.csv')


#df.loc[[37400]]
