==> Address/prajwal/Feb 20/mn 1850/lib/09_after running workflow.ipynb <==
import pandas as pd
import numpy as np

df1 = pd.read_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/mn-1850-final_tesseract.csv')
df2 = pd.read_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/mn-1850-3.csv')


df1.loc[2067]

list_of_empty_rows = list(df1[pd.isna(df1['hn_1'])==True][pd.isna(df1['best_match'])==True].index)

# Replace empty best_match values with those from df2
for a in (list_of_empty_rows):
    df1.at[a,'best_match'] = df2['Full Name'][a]

df1.loc[174]

# Change 255 255 Fulton to 255 Fulton
for a in range(0,len(df1)):
    try:
        hn1_to_search = int(df1['hn_1'][a])
        if df1['best_match'][a].index(str(hn1_to_search))>=0:
            #print(a, hn1_to_search, df1['best_match'][a])
            df1.at[a,'best_match'] = df1['best_match'][a].replace(str(hn1_to_search),'')
    except:
        pass

# change 255 Fulton & Court to 255 Fulton
for a in range(0,len(df1)):
    try:
        if df1['best_match'][a].index('&') >=0 and pd.isna(df1['hn_1'][a])==False:
            temp_str1 = df1['best_match'][a]
            loc_of_and = temp_str1.split().index('&')
            before_and = ' '.join(temp_str1.split()[0:loc_of_and])
            if len(before_and)>0:
                df1.at[a,'best_match'] = before_and
            else:
                df1.at[a,'best_match'] = ' '.join(temp_str1.split()[loc_of_and+1:])
    except:
        pass

# changes occurences like 'smith & nan' to 'smith'
for a in range(0, len(df1)):
    try:
        if df1['best_match'][a].index('nan &')>=0:
            print('1',a)
            df1.at[a,'best_match'] = df1['best_match'][a].replace('nan &','')
            print('1',df1['best_match'][a])
    except:
        pass
    try:
        if df1['best_match'][a].index('NAN &')>=0:
            print('2',a)
            print('2',df1['best_match'][a])
    except:
        pass
    try:
        if df1['best_match'][a].index('& nan')>=0:
            print('3',a)
            df1.at[a,'best_match'] = df1['best_match'][a].replace('& nan','')
            print('3',df1['best_match'][a])
    except:
        pass
    try:
        if df1['best_match'][a].index('& NAN')>=0:
            print('4',a)
            print('4',df1['best_match'][a])
    except:
        pass

df1.loc[2067]

#df1.to_csv('/Users/prajwal/Desktop/Columbia/C4SR/tesseract lib/mn 1850/data/mn_1850_for_rerun.csv')
